Cast registry value count to int in preview total
CountRegistryValues returns long; the preview total is an int and registry
value counts are small, so cast explicitly (CS0266).

+++ CleanMachine.Windows/WindowsCleanupService.cs
@@ -198,7 +198,7 @@ public sealed class WindowsCleanupService
                     break;
                 case CleanupKind.RegistryValues:
                     var count = CountRegistryValues(category.Path!);
-                    total += count;
+                    total += (int)count;
                     if (count > 0) shown.Add(new CleanupPreviewItem(category.Name, $"{count:N0} registry value(s) to clear", 0));
                     break;
                 case CleanupKind.RecycleBin:
